Created interface for IPAddressDetector and moved implementation into IPAddressDetectorImpl
